Refactor to create Referentializer.js + lint
