+++ src/serializer/utils.js
@@ -73,3 +73,11 @@ export function commonAncestorOf<T: HasParent>(node1: void | T, node2: void | T)
   }
   return n1;
 }
+
+// Gets map[key] with default value provided by defaultFn
+export function getOrDefault<K, V>(map: Map<K, V>, key: K, defaultFn: () => V): V {
+  let value = map.get(key);
+  if (!value) map.set(key, (value = defaultFn()));
+  invariant(value !== undefined);
+  return value;
+}
